+++ universum/lib/module_arguments.py
@@ -135,6 +135,16 @@ class ModuleArgumentParser(argparse.ArgumentParser):
         return True
 
     def _add_default_parser(self, args):
+        """
+        We need to manually add the default parser because of subcommand's arguments intersection.
+        1. We need subcommands for namespaces and help messages to be ajustable in different Universum modes.
+        2. If a subcommand is passed, and arguments in parser and subparser intersect, all values passed to
+           non-subparser are ignored; so without default subparser ``universum -vt=p4 submit`` will result in
+           undefined value of 'vcs_type', and ``universum -vt=p4 submit -vt=git`` will result in 'git' value.
+           Adding default subparser allows to avoid such unexpected behaviour.
+        3. Also if arguments of 'general' parser and a subparser intersect, the environment variables are
+           stored to 'general' parser (and then ignored as described above).
+        """
         subparsers_action = self._get_subparsers_action()
         if not subparsers_action:
             return