+++ universum/modules/reporter.py
@@ -1,12 +1,13 @@
 from collections import defaultdict
-from typing import Dict, List, Optional, TextIO, Tuple
+from typing import Dict, List, Tuple
+
 from typing_extensions import TypedDict
 
-from ..lib.gravity import Dependency
-from ..lib.utils import make_block
 from . import automation_server
 from .output import HasOutput
-from .structure_handler import HasStructure
+from .structure_handler import HasStructure, Block
+from ..lib.gravity import Dependency
+from ..lib.utils import make_block
 
 __all__ = [
     "ReportObserver",
@@ -48,6 +49,8 @@ class Reporter(HasOutput, HasStructure):
                             help="Send comment to review system on build success (in addition to vote up)")
         parser.add_argument("--report-only-fails", "-rof", action="store_true", dest="only_fails",
                             help="Include only the list of failed steps to reporting comments")
+        parser.add_argument("--report-only-fails-short", "-rofs", action="store_true", dest="only_fails_short",
+                            help="Include only the short list of failed steps to reporting comments")
         parser.add_argument("--report-no-vote", "-rnv", action="store_true", dest="no_vote",
                             help="Do not vote up/down review depending on result")
 
@@ -101,6 +104,9 @@ class Reporter(HasOutput, HasStructure):
             self.out.log("Not reporting: no build steps executed")
             return
 
+        if self.settings.only_fails_short:
+            self.settings.only_fails = True
+
         is_successful = True
         text = "Here is the summarized build result:\n"
         self.out.log(text)
@@ -143,16 +149,19 @@ class Reporter(HasOutput, HasStructure):
             for observer in self.observers:
                 observer.code_report_to_review(self.code_report_comments)
 
-    def _report_steps_recursively(self, block, text, indent):
+    def _report_steps_recursively(self, block: Block, text: str, indent: str) -> Tuple[str, bool]:
+        has_children: bool = bool(block.children)
+        block_title: str = block.number + ' ' + block.name
         if not self.settings.only_fails:
             text += indent + str(block) + '\n'
-            self.out.report_step(indent + str(block), block.status)
+            self.out.log_summary_step(indent + block_title, has_children, block.status)
         elif not block.is_successful():
-            text += str(block) + '\n'
-            self.out.report_step(str(block), block.status)
+            if not self.settings.only_fails_short or not has_children:
+                text += str(block) + '\n'
+                self.out.log_summary_step(block_title, has_children, block.status)
 
         is_successful = block.is_successful()
-        for substep in block.children:
-            text, status = self._report_steps_recursively(substep, text, indent + "  ")
+        for child in block.children:
+            text, status = self._report_steps_recursively(child, text, indent + "  ")
             is_successful = is_successful and status
         return text, is_successful